move jsdoc parser and tests to a separate folder and clean react parser
