+++ src/jsdoc/jsdoc/index.ts
@@ -1,0 +1,3 @@
+import parse from './jsdoc-parser';
+
+export default parse;
