+++ src/jsdoc/jsdoc/jsdoc-parser.ts
@@ -5,8 +5,6 @@ import { pathNormalizeToLinux } from '../../utils';
 import logger from '../../logger/logger';
 import { Doclet } from '../types';
 
-const reactDocs = require('react-docgen');
-
 function formatTag(tag: Record<string, any>): Record<string, any> {
   // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
   delete tag.title;
@@ -107,111 +105,29 @@ function extractDataRegex(doc: string, doclets: Array<Doclet>, filePath: PathOsB
   doclets.push(doclet);
 }
 
-function formatProperties(props) {
-  const parseDescription = description => {
-    // an extra step is needed to parse the properties description correctly. without this step
-    // it'd show the entire tag, e.g. `@property {propTypes.string} text - Button text.`
-    // instead of just `text - Button text.`.
-    try {
-      const descriptionAST = doctrine.parse(description, { unwrap: true, recoverable: true, sloppy: true });
-      if (descriptionAST && descriptionAST.tags[0]) return descriptionAST.tags[0].description;
-    } catch (err) {
-      // failed to parse the react property, that's fine, it'll return the original description
-    }
-    return description;
-  };
-  return Object.keys(props).map(name => {
-    const { type, description, required, defaultValue, flowType } = props[name];
-
-    return {
-      name,
-      description: parseDescription(description),
-      required,
-      type: stringifyType(type || flowType),
-      defaultValue
-    };
-  });
-}
-
-function formatMethods(methods) {
-  return Object.keys(methods).map(key => {
-    const { returns, modifiers, params, docblock, name } = methods[key];
-    return {
-      name,
-      description: docblock,
-      returns,
-      modifiers,
-      params
-    };
-  });
-}
-
-function fromReactDocs({ description, displayName, props, methods }, filePath): Doclet {
-  return {
-    filePath: pathNormalizeToLinux(filePath),
-    name: displayName,
-    description,
-    properties: formatProperties(props),
-    access: 'public',
-    // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
-    methods: formatMethods(methods)
-  };
-}
-
-function stringifyType(prop: { name: string; value?: any }): string {
-  const { name } = prop;
-  let transformed;
-
-  switch (name) {
-    default:
-      transformed = name;
-      break;
-    case 'func':
-      transformed = 'function';
-      break;
-    case 'shape':
-      transformed = JSON.stringify(
-        Object.keys(prop.value).reduce((acc = {}, current) => {
-          acc[current] = stringifyType(prop.value[current]);
-          return acc;
-        }, {})
-      );
-      break;
-    case 'enum':
-      transformed = prop.value.map(enumProp => enumProp.value).join(' | ');
-      break;
-    case 'instanceOf':
-      transformed = prop.value;
-      break;
-    case 'union':
-      transformed = prop.value.map(p => stringifyType(p)).join(' | ');
-      break;
-    case 'arrayOf':
-      transformed = `${stringifyType(prop.value)}[]`;
-      break;
-  }
-
-  return transformed;
-}
-
 export default async function parse(data: string, filePath?: PathOsBased): Promise<Doclet | []> {
   const doclets: Array<Doclet> = [];
   try {
-    const componentInfo = reactDocs.parse(data, undefined, undefined, { configFile: false });
-    if (componentInfo) {
-      const formatted = fromReactDocs(componentInfo, filePath);
-      formatted.args = [];
-      // this is a workaround to get the 'example' tag parsed when using react-docs
-      // because as of now Docgen doesn't parse @example tag, instead, it shows it inside
-      // the @description tag.
-      // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
-      extractDataRegex(formatted.description, doclets, filePath);
-      formatted.description = doclets[0].description;
-      formatted.examples = doclets[0].examples;
-      return formatted;
-    }
-  } catch (err) {
-    logger.debug(`failed parsing docs using docgen on path ${filePath} with error`, err);
+    /**
+     * [ \t]*  => can start with any number of tabs
+     * \/\*\*  => must start with exact `/**`
+     * \s*     => may follow with any number of white spaces
+     * [^*]*   => anything except star may repeat
+     * ([^\*]|(\*(?!\/)))* => may follow with stars as long as there is no "/" after the star
+     * \*\/ => must finish with a star and then \.
+     * This was taken as a combination of:
+     * https://stackoverflow.com/questions/35905181/regex-for-jsdoc-comments
+     * https://github.com/neogeek/jsdoc-regex/blob/master/index.js
+     */
+    const jsdocRegex = /[ \t]*\/\*\*\s*\n([^*]|(\*(?!\/)))*\*\//g;
+    const docs = data.match(jsdocRegex);
+    // populate doclets array
+    // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
+    // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
+    docs.forEach(doc => extractDataRegex(doc, doclets, filePath));
+  } catch (e) {
+    // never mind, ignore the doc of this source
+    logger.debug(`failed parsing docs using on path ${filePath} with error`, e);
   }
   // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
   return doclets.filter(doclet => doclet.access === 'public');
